+++ experimental/postcss-nesting/src/lib/merge-params.ts
@@ -1,0 +1,11 @@
+import { comma } from './list.js';
+
+export default function mergeParams(fromParams: string, toParams: string) {
+	return comma(fromParams)
+		.map((params1) =>
+			comma(toParams)
+				.map((params2) => `${params1} and ${params2}`)
+				.join(', '),
+		)
+		.join(', ');
+}