+++ plugin-packs/postcss-preset-env/src/plugins/plugins-by-id.mjs
@@ -37,6 +37,7 @@ import postcssReplaceOverflowWrap from 'postcss-replace-overflow-wrap';
 import postcssPlace from 'postcss-place';
 import postcssPrefersColorScheme from 'css-prefers-color-scheme';
 import postcssColorRebeccapurple from 'postcss-color-rebeccapurple';
+import postcssScopePseudoClass from '@csstools/postcss-scope-pseudo-class';
 import postcssSteppedValueFunctions from '@csstools/postcss-stepped-value-functions';
 import postcssFontFamilySystemUI from '../patch/postcss-system-ui-font-family.mjs';
 import postcssTextDecorationShorthand from '@csstools/postcss-text-decoration-shorthand';
@@ -85,6 +86,7 @@ export const pluginsById = new Map(
 		['place-properties', postcssPlace],
 		['prefers-color-scheme-query', postcssPrefersColorScheme],
 		['rebeccapurple-color', postcssColorRebeccapurple],
+		['scope-pseudo-class', postcssScopePseudoClass],
 		['stepped-value-functions', postcssSteppedValueFunctions],
 		['system-ui-font-family', postcssFontFamilySystemUI],
 		['text-decoration-shorthand', postcssTextDecorationShorthand],